+++ lib/polyfill/v2_2/enumerable.rb
@@ -1,6 +1,14 @@
 module Polyfill
   module V2_2
     module Enumerable
+      def max(n = nil)
+        return block_given? ? super(&::Proc.new) : super() if n.nil? || n == 1
+
+        raise ArgumentError, "negative size (#{n})" if n < 0
+
+        (block_given? ? sort(&::Proc.new) : sort).last(n).reverse
+      end
+
       def slice_after(pattern = nil)
         raise ArgumentError, 'both pattern and block are given' if pattern && block_given?
         raise ArgumentError, 'wrong number of arguments (given 0, expected 1)' if !pattern && !block_given?